move beanUtil into utils package. scope h2 dependency to test.
